feature(akita-filter-plugin): Add refresh method
Added akita-filter-plugin method, to force data refresh.
Will be more useful with the withServer feature, but also works with the local version.
Allows you to search the data in the store, and to pass the filters again. With the server version, it allows to force to call back the callback method of withServer, with the same data to call back the data from the backend (without changing anything of the filters (or the sorting, or the current page).
Akita-mat-datasource also has a method to call refresh from akita-filter-plugin.

Closes #50